Use 301 responses for redirects instead of meta refresh tags
When crawling the directory, if the file is an .html file that contains a meta refresh tag, configure the bucket object to be a website redirect that returns a 301 response when requested. We do this to allow #anchors to flow across to the redirect page, which doesn't work with the meta refresh tag -based redirects.